Fix build for multi-arch macOS binaries
For example, in order to be compatible with both x86-64 and new
Apple M1, serf would be compiled with:
  CFLAGS   ="-arch x86_64 -arch arm64"
  CPPFLAGS ="-arch x86_64 -arch arm64"
  LINKFLAGS="-arch x86_64 -arch arm64"

Counter-intuitively, when scons 'MergeFlags()' is called with
'unique=1', it also affects already present flags and not just the
merged ones. This causes flags to become 'x86_64 -arch arm64', where
the first '-arch' is now gone, which in turn causes build error
  error: no such file or directory: 'x86_64'

Fix that by passing 'unique=0'. In my case, it was sufficient to fix
in 'parse_libs()' only, but I decided to apply it to the other
'ParseConfig()' as well. See also r1712172, which fixed SERF-160. The
bugreport there is not very detailed, but it also involved breaking
linker flags.

* SConstruct
  When merging additional build flags, use 'unique=0' to avoid
  corrupting existing flags.